+++ packages/core/src/utils/pluginRenderElement.tsx
@@ -15,7 +15,13 @@ import { getRenderNodeProps } from './getRenderNodeProps';
  */
 export const pluginRenderElement = <V extends Value>(
   editor: PlateEditor<V>,
-  { key, type, component: _component, props }: PlatePlugin<{}, V>
+  {
+    key,
+    type,
+    component: _component,
+    dangerouslyAllowAttributes,
+    props,
+  }: PlatePlugin<{}, V>
 ): RenderElement =>
   function render(nodeProps) {
     const { element, children: _children } = nodeProps;
@@ -32,6 +38,7 @@ export const pluginRenderElement = <V extends Value>(
 
       nodeProps = getRenderNodeProps({
         attributes: element.attributes as any,
+        dangerouslyAllowAttributes,
         nodeProps: nodeProps as any,
         props,
         type: type!,