+++ packages/core/src/utils/pluginRenderLeaf.tsx
@@ -14,7 +14,13 @@ import { getRenderNodeProps } from './getRenderNodeProps';
  */
 export const pluginRenderLeaf = <V extends Value>(
   editor: PlateEditor<V>,
-  { key, type = key, component, props }: PlatePlugin<{}, V>
+  {
+    key,
+    type = key,
+    component,
+    dangerouslyAllowAttributes,
+    props,
+  }: PlatePlugin<{}, V>
 ): RenderLeaf =>
   function render(nodeProps) {
     const { leaf, children } = nodeProps;
@@ -24,6 +30,7 @@ export const pluginRenderLeaf = <V extends Value>(
 
       nodeProps = getRenderNodeProps({
         attributes: leaf.attributes as any,
+        dangerouslyAllowAttributes,
         props,
         nodeProps: nodeProps as any,
         type,