+++ packages/media/src/media-embed/createMediaEmbedPlugin.ts
@@ -13,6 +13,7 @@ export const createMediaEmbedPlugin = createPluginFactory<MediaPlugin>({
   key: ELEMENT_MEDIA_EMBED,
   isElement: true,
   isVoid: true,
+  dangerouslyAllowAttributes: ['width', 'height'],
   options: {
     transformUrl: parseIframeUrl,
   },